Web design, connected addcontacts and registration to database.
Web app design was finished for log in and Sign
Up page, both of these are connected to the database.
Designed user account log in page and added functional
buttons to search and add contacts. The AddContacts
feature is fully functional and connected to the database as well.
Php and jason files were also modified to provide
functionality.
All changes were uploaded to the server through file zilla.
The website is live and accessible.